+++ idle-backend/src/main/java/com/idle/mapper/OrdersMapper.java
@@ -12,15 +12,15 @@ import java.util.List;
 public interface OrdersMapper {
     @Insert("insert into db_orders (id, create_time, seller, buyer, payment, price, shipping_address) values " +
             "(#{id}, #{create_time}, #{seller}, #{buyer}, #{payment}, #{price}, #{shipping_address})")
-    int createOrder(Integer id, String create_time, String seller, String buyer, String payment, BigDecimal price,
+    int createOrder(Long id, String create_time, String seller, String buyer, String payment, BigDecimal price,
                     String shipping_address);
 
     @Select("select * from db_orders where #{nickname}=db_orders.seller or #{nickname}=db_orders.buyer")
     List<Orders> getUserOrders(String nickname);
 
     @Delete("delete * from db_orders where id = #{id}")
-    int deleteOrder(Integer id);
+    int deleteOrder(Long id);
 
     @Select("select * from db_orders where id = #{id} and (db_orders.buyer = #{nickname} or db_orders.seller = #{nickname})")
-    List<Orders> getOrderById(Integer id, String nickname);
+    List<Orders> getOrderById(Long id, String nickname);
 }
