+++ idle-backend/src/main/java/com/idle/service/GoodsService.java
@@ -8,25 +8,25 @@ import java.util.List;
 
 public interface GoodsService {
 
-    List<Goods> findGoodsById(Integer id) throws Exception;
+    List<Goods> findGoodsById(Long id) throws Exception;
 
-    boolean createGoods(Integer id, Integer publisher_id, String product_name, BigDecimal price, String category,
+    boolean createGoods(Long id, Long publisher_id, String product_name, BigDecimal price, String category,
                       String introduce, String picture, String create_time);
 
-    boolean updateGoods(Integer id, String product_name, BigDecimal price, String category,
+    boolean updateGoods(Long id, String product_name, BigDecimal price, String category,
                       String introduce, String picture);
 
-    boolean updateStatus(Integer id);
+    boolean updateStatus(Long id);
 
     List<Goods> getAllGoods();
 
-    List<Goods> getGoodsByPublisherId(Integer publisher_id);
+    List<Goods> getGoodsByPublisherId(Long publisher_id);
 
-    AccountUser getUserByPublisherId(Integer publisher_id);
+    AccountUser getUserByPublisherId(Long publisher_id);
 
     List<Goods> getGoodsByCategory(String category);
 
     List<Goods> getGoodsByProductNameOrIntroduce(String search);
 
-    String getPriceById(Integer id);
+    String getPriceById(Long id);
 }
